+++ src/main/java/com/example/calendar/Model/Participant.java
@@ -19,7 +19,7 @@ import java.util.Set;
 @Setter
 @EqualsAndHashCode
 @ToString
-public class Participant implements UserDetails {
+public class Participant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -55,38 +55,5 @@ public class Participant implements UserDetails {
         }
     }
 
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Set.of(getRole());
-    }
 
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return getStatus().equals(ParticipantStatus.ACTIVE);
-    }
 }
